Rename notes/common to notes/pkg/utils
